Change key_state from float to int. Remove key_state_tracker::zero().
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@479 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/input/key_type.hpp
@@ -311,7 +311,7 @@ inline bool is_alt(const key_code key)
 		(key == kc::key_altgr);
 }
 
-typedef float key_state;
+typedef int key_state;
 
 class key_pair {
 public:
